Fix use of double-quoted strings in SQLite queries
Double-quotes are used for identifiers, not strings. While SQLite will interpret unknown identifiers that use double-quotes as strings, this is considered a misfeature. This commit fixes the SQL for those of us with double-quoted strings disabled, as well as fixes setting domain_name to the name of a column in the SQLite table.